Usa zip existente se possível

+++ cidades.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 
+import os
 import urllib.request
 import zipfile
 
@@ -8,9 +9,12 @@ url = 'http://download.geonames.org/export/zip/BR.zip'
 zip_path = '/tmp/BR.zip'
 unzip_path = '/tmp/cidades/'
 
-# Download das cidades brasileiras
-print('Baixando cidades brasileiras (BR.zip)')
-urllib.request.urlretrieve(url, zip_path)
+if not os.path.exists(zip_path):
+    # Download das cidades brasileiras
+    print('Baixando cidades brasileiras (BR.zip)')
+    urllib.request.urlretrieve(url, zip_path)
+else:
+    print('Arquivo de cidades brasileiras já existe')
 
 # Descompacta zip das cidades
 with zipfile.ZipFile(zip_path) as cidades_zip:
